Add option for adding audio narration to posts, with shortcut.
Putting a 'narration' key in the front-matter with a source
path / URL will cause an audio player to show up, informing
the reader that there is narration available. This can be
toggled with the 'p' key.

This took some re-writing of the stylesheet to get margins and
spacing to look correct with any combination of chips and
audio, since the two are now in the same "extras" div.