Layout change: Properly align <th> and <td> text.
I genuinely have NO idea where the CSS rule that sets *everything*
to left-align is, but it's not in the Sass included with this, I
know that much. Best I can do is that hack-job, and manually
override the text direction for each possible value that Markdown
generates, since just slapping 'text-align: unset;' or 'inherit',
or 'initial' over it does nothing at all.